Refactor AutoStage: split into focused files, new namespace
Split AutoStage.cs (281 lines) into 3 focused files:
- AutoStageState.cs: AfcAutoStage enum, state, enum injection, ApplyPatches
- GaugeButtonPatches.cs: ToggleEnum, IsSet, IsFlightComputerDisabled
- AutoStageExecution.cs: propellant depletion detection + staging

Namespace: Features -> Features.AutoStage
